Move token checks to core module, use in payouts
